Refactor de sellado de placas y actualizar registroSP
